SGF-373 - Implement a Spring Session Adapter for GemFire to back an HttpSession similar to Redis.
Refactor the GemfireOperationsSessionRepository.GemFireSession class to implement the Comparable<ExpiringSession> interface.
Modify the sessionAttributes Map field in GemFireSessionAttributes to be transient in order to prevent explicit serialization of this field when using Reflection-based auto-serialization in GemFire.
Mark the 'id' property as the identity field in the PDX serialized instance created in the toData(..) method.
Refactor several unit and integration test class methods.